Add first set of data files
Data files obtained with a 400dpi mouse on a T440s, see below for the
libpointing output for this device. CPI specified on commandline.

The three transfer functions are osx:, windows: and xorg: in their default
settings. Note that xorg has a *lot* of gain mismatch warnings, the validity
of that file is to be questioned.

Data produced by moving the mouse at various speeds for a while. Max speeds
were wildly moving the mouse around, beyond what is sensible in desktop
interaction.

*** NOTE ***
Data files produced before normalizing input, deltas are in CPI400 here
*** NOTE ***

Data files processed with:
   cat datapoints.dat | sed "/.* 0\.000000/d" | sort -g | uniq -f 1

Pointing device
  input:/dev/input/event7?debugLevel=0&cpi=400&hz=100&seize=0
  400 CPI, 100 Hz
  device is active

Display device
  xorgdisplay::0
  1920 x 1200 pixels, 508 x 318 mm
  96 x 95.8491 PPI, 60 Hz